move: events dialog and non dialog files
