+++ frontend/app/(settings)/profile/registrations/page.tsx
@@ -13,7 +13,7 @@ import {useCallback, useEffect, useState} from "react";
 import {getClient} from "@/lib/graphql";
 import {useUser} from "@/components/providers";
 import PlannerItem from "@/components/planner-item";
-import {EventDialog} from "@/components/event-calendar/event-dialog";
+import {EventDialog} from "@/components/dialog/events/event-dialog";
 import {defaultEvent} from "@/types/defaults";
 
 import {Accordion, AccordionContent, AccordionItem, AccordionTrigger,} from "@/components/ui/accordion";
