+++ frontend/components/event-calendar/index.ts
@@ -3,7 +3,7 @@ export { AgendaView } from "./agenda-view"
 export { DayView } from "./day-view"
 export { CalendarEventItem } from "./calendar-event-item"
 export { CalendarCell } from "./calendar-cell"
-export { EventDialog } from "./event-dialog"
+export { EventDialog } from "../dialog/events/event-dialog"
 export { EventItem } from "./event-item"
 export { EventsPopup } from "./events-popup"
 export { EventCalendar } from "./event-calendar"
