+++ frontend/components/tutorial-element.tsx
